+++ ds/go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/iotaledger/hive.go/constraints v0.0.0-20230630101126-a9b43f69f1f0
 	github.com/iotaledger/hive.go/ierrors v0.0.0-20230630101126-a9b43f69f1f0
 	github.com/iotaledger/hive.go/lo v0.0.0-20230630101126-a9b43f69f1f0
+	github.com/iotaledger/hive.go/runtime v0.0.0-20230629181801-64c530ff9d15
 	github.com/iotaledger/hive.go/serializer/v2 v2.0.0-rc.1.0.20230630101126-a9b43f69f1f0
 	github.com/iotaledger/hive.go/stringify v0.0.0-20230630101126-a9b43f69f1f0
 	github.com/mr-tron/base58 v1.2.0
