+++ ds/set/set.go
@@ -1,18 +1,12 @@
 package set
 
 import (
-	"fmt"
-	"reflect"
-	"strings"
+	"sync"
 
 	"github.com/iotaledger/hive.go/ds/orderedmap"
 	"github.com/iotaledger/hive.go/ds/types"
-	"github.com/iotaledger/hive.go/ds/walker"
-	"github.com/iotaledger/hive.go/ierrors"
 	"github.com/iotaledger/hive.go/lo"
 )
-
-// region Set
 
 // Set is a generic thread-safe collection of unique elements.
 type Set[ElementType comparable] interface {
@@ -28,6 +22,9 @@ type Set[ElementType comparable] interface {
 	// DeleteAll deletes the given elements from the set.
 	DeleteAll(other ReadOnly[ElementType]) (removedElements Set[ElementType])
 
+	// Apply tries to apply the given mutations to the set atomically and returns the applied mutations.
+	Apply(mutations Mutations[ElementType]) (appliedMutations Mutations[ElementType])
+
 	// Decode decodes the set from a byte slice.
 	Decode(b []byte) (bytesRead int, err error)
 
@@ -40,9 +37,11 @@ type Set[ElementType comparable] interface {
 
 // New creates a new Set with the given elements.
 func New[T comparable](elements ...T) Set[T] {
-	s := &set[T]{&readOnly[T]{
-		OrderedMap: orderedmap.New[T, types.Empty](),
-	}}
+	s := &set[T]{
+		readOnly: &readOnly[T]{
+			OrderedMap: orderedmap.New[T, types.Empty](),
+		},
+	}
 
 	for _, element := range elements {
 		s.OrderedMap.Set(element, types.Void)
@@ -54,15 +53,23 @@ func New[T comparable](elements ...T) Set[T] {
 // set implements the Set interface.
 type set[ElementType comparable] struct {
 	*readOnly[ElementType]
+
+	mutex sync.RWMutex
 }
 
 // Add adds a new element to the set and returns true if the element was not present in the set before.
 func (s *set[ElementType]) Add(element ElementType) bool {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
 	return !lo.Return2(s.Set(element, types.Void))
 }
 
-// AddAll adds all elements to the set and returns true if any element has been added.
+// AddAll tries to add all elements to the set and returns the elements that have been added.
 func (s *set[ElementType]) AddAll(elements ReadOnly[ElementType]) (addedElements Set[ElementType]) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
 	addedElements = New[ElementType]()
 	_ = elements.ForEach(func(element ElementType) (err error) {
 		if !lo.Return2(s.Set(element, types.Void)) {
@@ -77,11 +84,17 @@ func (s *set[ElementType]) AddAll(elements ReadOnly[ElementType]) (addedElements
 
 // Delete deletes the given element from the set.
 func (s *set[ElementType]) Delete(element ElementType) bool {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
 	return s.OrderedMap.Delete(element)
 }
 
 // DeleteAll deletes the given elements from the set.
 func (s *set[ElementType]) DeleteAll(other ReadOnly[ElementType]) (removedElements Set[ElementType]) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
 	removedElements = New[ElementType]()
 	_ = other.ForEach(func(element ElementType) (err error) {
 		if s.Delete(element) {
@@ -94,6 +107,28 @@ func (s *set[ElementType]) DeleteAll(other ReadOnly[ElementType]) (removedElemen
 	return removedElements
 }
 
+// Apply tries to apply the given mutations to the set atomically and returns the applied mutations.
+func (s *set[ElementType]) Apply(mutations Mutations[ElementType]) (appliedMutations Mutations[ElementType]) {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+
+	addedElements := New[ElementType]()
+	mutations.AddedElements().Range(func(element ElementType) {
+		if !lo.Return2(s.Set(element, types.Void)) {
+			addedElements.Add(element)
+		}
+	})
+
+	removedElements := New[ElementType]()
+	mutations.DeletedElements().Range(func(element ElementType) {
+		if s.OrderedMap.Delete(element) {
+			removedElements.Add(element)
+		}
+	})
+
+	return NewMutations[ElementType]().WithAddedElements(addedElements).WithDeletedElements(removedElements)
+}
+
 // ToReadOnly returns a read-only version of the set.
 func (s *set[ElementType]) ToReadOnly() ReadOnly[ElementType] {
 	return s.readOnly
@@ -101,180 +136,3 @@ func (s *set[ElementType]) ToReadOnly() ReadOnly[ElementType] {
 
 // code contract (make sure the type implements all required methods).
 var _ Set[int] = new(set[int])
-
-// endregion
-
-// region ReadOnly
-
-// ReadOnly is a generic thread-safe collection of unique elements that can only be read.
-type ReadOnly[ElementType comparable] interface {
-	// Has returns true if the set contains the given element.
-	Has(key ElementType) (has bool)
-
-	// HasAll returns true if the set contains all elements of the given set.
-	HasAll(other ReadOnly[ElementType]) bool
-
-	// ForEach iterates through all elements of the set (returning an error will stop the iteration).
-	ForEach(callback func(element ElementType) error) error
-
-	// Range iterates through all elements of the set.
-	Range(callback func(element ElementType))
-
-	// Intersect returns the intersection of the set and the given set.
-	Intersect(other ReadOnly[ElementType]) Set[ElementType]
-
-	// Filter returns a new set with all elements that satisfy the given predicate.
-	Filter(predicate func(element ElementType) bool) Set[ElementType]
-
-	// Equals returns true if the set contains the same elements as the given set.
-	Equals(other ReadOnly[ElementType]) bool
-
-	// Is returns true if the given element is the only element in the set.
-	Is(element ElementType) bool
-
-	// Iterator returns an iterator for the set.
-	Iterator() *walker.Walker[ElementType]
-
-	// Clone returns a shallow copy of the set.
-	Clone() Set[ElementType]
-
-	// Size returns the number of elements in the set.
-	Size() int
-
-	// IsEmpty returns true if the set is empty.
-	IsEmpty() bool
-
-	// Clear removes all elements from the set.
-	Clear()
-
-	// ToSlice returns a slice representation of the set.
-	ToSlice() []ElementType
-
-	// Encode encodes the set into a byte slice.
-	Encode() ([]byte, error)
-
-	// String returns a string representation of the set.
-	String() string
-}
-
-// readOnly implements the ReadOnly interface.
-type readOnly[T comparable] struct {
-	*orderedmap.OrderedMap[T, types.Empty] `serix:"0"`
-}
-
-// HasAll returns true if the set contains all elements of the given set.
-func (r *readOnly[T]) HasAll(other ReadOnly[T]) bool {
-	if r == nil {
-		return false
-	}
-
-	return other.ForEach(func(element T) error {
-		if !r.Has(element) {
-			return ierrors.New("element not found")
-		}
-
-		return nil
-	}) == nil
-}
-
-// ForEach iterates through all elements of the set (returning an error will stop the iteration).
-func (r *readOnly[T]) ForEach(callback func(element T) error) (err error) {
-	if r == nil {
-		return nil
-	}
-
-	r.OrderedMap.ForEach(func(element T, _ types.Empty) bool {
-		if err = callback(element); err != nil {
-			return false
-		}
-
-		return true
-	})
-
-	return err
-}
-
-// Range iterates through all elements of the set.
-func (r *readOnly[T]) Range(callback func(element T)) {
-	if r != nil {
-		r.OrderedMap.ForEach(func(element T, _ types.Empty) bool {
-			callback(element)
-
-			return true
-		})
-	}
-}
-
-// Intersect returns the intersection of the set and the given set.
-func (r *readOnly[T]) Intersect(other ReadOnly[T]) (intersection Set[T]) {
-	return r.Filter(other.Has)
-}
-
-// Filter returns a new set with all elements that satisfy the given predicate.
-func (r *readOnly[T]) Filter(predicate func(element T) bool) (filtered Set[T]) {
-	filtered = New[T]()
-	_ = r.ForEach(func(element T) (err error) {
-		if predicate(element) {
-			filtered.Add(element)
-		}
-
-		return nil
-	})
-
-	return filtered
-}
-
-// Equals returns true if the set contains the same elements as the given set.
-func (r *readOnly[T]) Equals(other ReadOnly[T]) (equal bool) {
-	return r == other || (r != nil && other != nil && r.Size() == other.Size() && r.HasAll(other))
-}
-
-// Is returns true if the given element is the only element in the set.
-func (r *readOnly[T]) Is(element T) bool {
-	return r.Size() == 1 && r.Has(element)
-}
-
-// Iterator returns an iterator for the set.
-func (r *readOnly[T]) Iterator() *walker.Walker[T] {
-	return walker.New[T](false).PushAll(r.ToSlice()...)
-}
-
-// Clone returns a shallow copy of the set.
-func (r *readOnly[T]) Clone() (cloned Set[T]) {
-	return New[T]().AddAll(r)
-}
-
-// ToSlice returns a slice representation of the set.
-func (r *readOnly[T]) ToSlice() (slice []T) {
-	slice = make([]T, 0)
-
-	if r != nil {
-		_ = r.ForEach(func(element T) error {
-			slice = append(slice, element)
-
-			return nil
-		})
-	}
-
-	return slice
-}
-
-// String returns a string representation of the set.
-func (r *readOnly[T]) String() (humanReadable string) {
-	var elementType T
-	elementTypeName := reflect.TypeOf(elementType).Name()
-
-	elementStrings := make([]string, 0)
-	_ = r.ForEach(func(element T) (err error) {
-		elementStrings = append(elementStrings, strings.TrimRight(strings.ReplaceAll(fmt.Sprintf("%+v", element), elementTypeName+"(", ""), ")"))
-
-		return nil
-	})
-
-	return fmt.Sprintf("%ss(%s)", elementTypeName, strings.Join(elementStrings, ", "))
-}
-
-// code contract (make sure the type implements all required methods).
-var _ ReadOnly[int] = new(readOnly[int])
-
-// endregion
